started working on analyze m.m
