+++ Documentation/Rapport/Main.ist
@@ -1,5 +1,5 @@
 % makeindex style file created by the glossaries package
-% for document 'Main' on 2022-10-18
+% for document 'Main' on 2022-10-19
 actual '?'
 encap '|'
 level '!'
