+++ Documentation/Rapport/Main.tex
@@ -287,7 +287,7 @@ keepaspectratio]{background-1.png}
       description={Enigma var ett krypterings verktyg som användes under andra världskriget av tyska
       militären för att kryptera medelanden. Maskinen bestod av en elektromekanisk rotordisk som under tiden
       medellandet skrivs in för kryptering även ändrar de elektriska kopplingarna mellan vilka bokstäver som blir vad.
-      Detta är en av sakerna som gjorde enigma väldigt svår att knäcka och en av anledningarna till att
+      Detta är en av sakerna som gjorde enigma väldigt svår att knäcka samt en av anledningarna till att
       liknande maskiner användes under stora delar av de tidiga 1900-talet.\cite{enigma-ref}}
     }
