+++ tests/Analyze.py
@@ -1,19 +1,23 @@
 from PyAES.encrypt import encrypt
 from PyAES.decrypt import decrypt
+from Analyze_init import setup, clean
+from timeit import timeit
+from datetime import timedelta
 
 
-def main(i):
-    key = "2b7e151628aed2a6abf7158809cf4f3c"
-    iv = "000102030405060708090a0b0c0d0e0f"
-    running_mode = "ECB"
-    file_path = r"/Users/gabriellindeblad/Documents/GitHub/AES-Python/tmp/test_files/sword_fifth_bakground_light_1.jpg"
+def main():
+    # Initialize and create test data
+    setup()
 
-    if i == "enc":
-        encrypt(key, file_path, running_mode)
-    else:
-        file_path += ".enc"
-        decrypt(key, file_path, running_mode)
+    # Preform analysis tests
+    timeit.d
+
+    # Display time
+    print(timedelta(seconds=start - end))
+
+    # Cleaning up after test
+    clean()
 
 
 if __name__ == '__main__':
-    main("enc")
+    main()
